19/11/2024 "El barco va bien encaminado"

+++ Bulonera/src/main/java/Bulonera/logica/cuenta_corriente.java
@@ -27,7 +27,7 @@ public class cuenta_corriente implements Serializable {
     @GeneratedValue(strategy=GenerationType.AUTO)
     private int id_cc;
     @Basic
-    private int debe_cc, haber_cc, saldo_cc;
+    private double debe_cc, haber_cc, saldo_cc;
     @Temporal(TemporalType.DATE)
     private Date fecha_cc;
     
@@ -40,7 +40,7 @@ public class cuenta_corriente implements Serializable {
     {
     }
 
-    public cuenta_corriente(int id_cc, int debe_cc, int haber_cc, int saldo_cc, Date fecha_cc, ArrayList<pago> listaPagos_cc, cabecera_remito cabeceraremito) {
+    public cuenta_corriente(int id_cc, double debe_cc, double haber_cc, double saldo_cc, Date fecha_cc, ArrayList<pago> listaPagos_cc, cabecera_remito cabeceraremito) {
         this.id_cc = id_cc;
         this.debe_cc = debe_cc;
         this.haber_cc = haber_cc;
@@ -74,27 +74,27 @@ public class cuenta_corriente implements Serializable {
         this.id_cc = id_cc;
     }
 
-    public int getDebe_cc() {
+    public double getDebe_cc() {
         return debe_cc;
     }
 
-    public void setDebe_cc(int debe_cc) {
+    public void setDebe_cc(double debe_cc) {
         this.debe_cc = debe_cc;
     }
 
-    public int getHaber_cc() {
+    public double getHaber_cc() {
         return haber_cc;
     }
 
-    public void setHaber_cc(int haber_cc) {
+    public void setHaber_cc(double haber_cc) {
         this.haber_cc = haber_cc;
     }
 
-    public int getSaldo_cc() {
+    public double getSaldo_cc() {
         return saldo_cc;
     }
 
-    public void setSaldo_cc(int saldo_cc) {
+    public void setSaldo_cc(double saldo_cc) {
         this.saldo_cc = saldo_cc;
     }
